GCF file split (munge): slack
